refatoracao da estrutura com req 11 concluido :)
