+++ src/App.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { Home, Cart } from './pages';
-import ProductDetails from './pages/ProductDetails';
+import { Home, Cart, Details } from './pages';
 
 function App() {
   return (
@@ -10,7 +9,7 @@ function App() {
       <Switch>
         <Route exact path="/" component={ Home } />
         <Route exact path="/cart" component={ Cart } />
-        <Route exact path="/:id" render={ (props) => <ProductDetails { ...props } /> } />
+        <Route exact path="/:id" render={ (props) => <Details { ...props } /> } />
       </Switch>
     </BrowserRouter>
   );
