+++ src/components/CardProduct/index.jsx
@@ -1,9 +1,37 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import ButtonAdd from './ButtonAdd';
 
 class ProductCard extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      disabled: false,
+    };
+  }
+
+  handleClick = () => {
+    const { product } = this.props;
+    const cartList = JSON.parse(localStorage.getItem('cartList')) || [];
+    cartList.push(product);
+    localStorage.setItem('cartList', JSON.stringify(cartList));
+    this.setState({ disabled: true });
+  }
+
+  buttonAdd = () => {
+    const { disabled } = this.state;
+    return (
+      <button
+        data-testid="product-add-to-cart"
+        onClick={ this.handleClick }
+        type="button"
+        disabled={ disabled }
+      >
+        Adicionar ao carrinho
+      </button>
+    );
+  }
+
   render() {
     const { product: { id, title, thumbnail, price } } = this.props;
     return (
@@ -13,7 +41,7 @@ class ProductCard extends React.Component {
           <img src={ thumbnail } alt="" />
           {price}
         </Link>
-        <ButtonAdd id={ id } testId="product-add-to-cart">Add</ButtonAdd>
+        {this.buttonAdd()}
       </div>
     );
   }
