--- src/components/Coments.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Coments extends React.Component {
-  render() {
-    const objs = localStorage.getItem('comentario');
-    return (
-      <div>
-        {console.log(objs)}
-      </div>
-    );
-  }
-}
-
-export default Coments;
