+++ src/components/index.js
@@ -1,6 +1,5 @@
-export { default as CategoriesList } from './CategoriesList';
-export { default as ProductCard } from './ProductCard';
-export { default as ProductList } from './ProductList';
-export { default as SearchBar } from './SearchBar';
-export { default as Form } from './Form';
-export { default as ProductCart } from './ProductCart';
+export { default as CardProduct } from './CardProduct';
+export { default as CartProduct } from './CartProduct';
+export { default as Comments } from './Comments';
+
+export { getCategories, getProductsFromCategoryAndQuery } from '../services/api';
