+++ src/pages/index.js
@@ -1,2 +1,3 @@
 export { default as Home } from './Home';
 export { default as Cart } from './Cart';
+export { default as Details } from './Details';
